+++ src/Events.fs
@@ -39,6 +39,19 @@ type Resource = {
     Href: string
 }
 
+type ResourceDto = {
+    name: string
+    href: string
+}
+
+[<RequireQualifiedAccess>]
+module Resource =
+    let toDto (resource: Resource) =
+        {
+            name = resource.Name
+            href = resource.Href
+        }
+
 // Generic event
 
 type Event<'KeyData, 'MetaData, 'DomainData> = {
@@ -76,6 +89,36 @@ type CommonEvent = {
     Resource: Resource option
 }
 
+type NoData = unit
+
+type EventDto<'ResourceDto, 'KeyDataDto, 'MetaDataDto, 'DomainDataDto> = {
+    schema: int
+    id: Guid
+    correlation_id: Guid
+    causation_id: Guid
+    timestamp: string
+    event: string
+    domain: string
+    context: string
+    purpose: string
+    version: string
+    zone: string
+    bucket: string
+    resource: 'ResourceDto
+    meta_data: 'MetaDataDto
+    key_data: 'KeyDataDto
+    domain_data: 'DomainDataDto
+}
+
+type SerializeEvent<'KeyData, 'MetaData, 'DomainData, 'ResourceDto, 'KeyDataDto, 'MetaDataDto, 'DomainDataDto, 'Error> =
+    (Event<'KeyData, 'MetaData, 'DomainData> -> Result<unit, 'Error>)
+        -> (Resource option -> Result<'ResourceDto, 'Error>)
+        -> ('MetaData -> Result<'MetaDataDto, 'Error>)
+        -> ('KeyData -> Result<'KeyDataDto, 'Error>)
+        -> ('DomainData -> Result<'DomainDataDto, 'Error>)
+        -> Event<'KeyData, 'MetaData, 'DomainData>
+        -> Result<EventDto<'ResourceDto, 'KeyDataDto, 'MetaDataDto, 'DomainDataDto>, 'Error>
+
 [<RequireQualifiedAccess>]
 module Event =
     let toCommon (event: Event<'KeyData, 'MetaData, 'DomainData>) =
@@ -95,6 +138,36 @@ module Event =
             Resource = event.Resource
         }
 
+    let toDto: SerializeEvent<'KeyData, 'MetaData, 'DomainData, 'ResourceDto, 'KeyDataDto, 'MetaDataDto, 'DomainDataDto, 'Error> =
+        fun assertEventType serializeResource serializeMetaData serializeKeyData serializeDomainData event ->
+            result {
+                do! assertEventType event
+
+                let! resourceDto = event.Resource |> serializeResource
+                let! metaDataDto = event.MetaData |> serializeMetaData
+                let! keyDataDto = event.KeyData |> serializeKeyData
+                let! domainDataDto = event.DomainData |> serializeDomainData
+
+                return {
+                    schema = event.Schema
+                    id = event.Id |> EventId.value
+                    correlation_id = event.CorrelationId |> CorrelationId.value
+                    causation_id = event.CausationId |> CausationId.value
+                    timestamp = event.Timestamp
+                    event = event.Event |> EventName.value
+                    domain = event.Domain |> Domain.value
+                    context = event.Context |> Context.value
+                    purpose = event.Purpose |> Purpose.value
+                    version = event.Version |> Version.value
+                    zone = event.Zone |> Zone.value
+                    bucket = event.Bucket |> Bucket.value
+                    meta_data = metaDataDto
+                    resource = resourceDto
+                    key_data = keyDataDto
+                    domain_data = domainDataDto
+                }
+            }
+
 [<RequireQualifiedAccess>]
 module CommonEvent =
     let box (event: CommonEvent) =
@@ -107,6 +180,13 @@ module CommonEvent =
             event.Bucket
 
 type RawData = RawData of FSharp.Data.JsonValue
+
+[<RequireQualifiedAccess>]
+module RawData =
+    open FSharp.Data
+
+    let toJson (RawData data) =
+        data.ToString(JsonSaveOptions.DisableFormatting)
 
 type RawEvent = Event<RawData, RawData option, RawData option>